# nextjs-build/src/frontend/TestFramework.Client.WebAPI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TestFramework.WebAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/UIElements/{id}": {
      "get": {
        "tags": [
          "UIElements"
        ],
        "operationId": "GetUIElementById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIElementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIElementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIElementDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "UIElements"
        ],
        "operationId": "PutUIElement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIElemenDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIElemenDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIElemenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UIElements"
        ],
        "operationId": "DeleteUIElement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/UIElements": {
      "get": {
        "tags": [
          "UIElements"
        ],
        "operationId": "GetUIElementPage",
        "parameters": [
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIElementDtoPaginationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIElementDtoPaginationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIElementDtoPaginationResult"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UIElements"
        ],
        "operationId": "PostUIElement",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIElemenDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIElemenDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIElemenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/UIPages/{id}": {
      "get": {
        "tags": [
          "UIPages"
        ],
        "operationId": "GetUIPageById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "UIPages"
        ],
        "operationId": "PutUIPage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIPageDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIPageDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIPageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UIPages"
        ],
        "operationId": "DeleteUIPage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/UIPages": {
      "get": {
        "tags": [
          "UIPages"
        ],
        "operationId": "GetUIPage",
        "parameters": [
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDtoPaginationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDtoPaginationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDtoPaginationResult"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UIPages"
        ],
        "operationId": "PostUIPage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIPageDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIPageDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrCreateUIPageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/UIPages/{id}/UIElements": {
      "get": {
        "tags": [
          "UIPages"
        ],
        "operationId": "GetUIPageDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDetailsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDetailsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUIPageDetailsDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeOrCreateUIElemenDto": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string",
            "nullable": true
          },
          "FindByMethod": {
            "enum": [
              "ById",
              "ByClass",
              "ByXPath"
            ],
            "type": "string"
          },
          "FindBy": {
            "type": "string",
            "nullable": true
          },
          "PageId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ChangeOrCreateUIPageDto": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorCodesResult": {
        "type": "object",
        "properties": {
          "ErrorCodes": {
            "type": "array",
            "items": {
              "enum": [
                "IdLessThen1",
                "UIElements_NameMaxLength",
                "UIElements_NameIsEmpty",
                "UIElements_NameAlreadyUsed",
                "UIElements_FindByIsEmpty",
                "UIElements_FindByMaxLength",
                "UITestCases_NameMaxLength",
                "UITestCases_NameIsEmpty",
                "UITestCases_NameAreadyUsed",
                "UITestCases_StartUrlIsEmpty",
                "UITestCases_UIElementNotFound",
                "UITestCases_UserFilesNotFound",
                "UITestCases_StepsAreNotUnique",
                "UIEvents_NameIsEmtpy",
                "UIEvents_NameMaxLength",
                "UITestRuns_TestCasesNotFound",
                "UITestRuns_TestCasesAreEmpty",
                "UIPages_NameIsEmpty",
                "UIPages_NameMaxLength",
                "UITestCaseStatusInfo_StartIsEmpty",
                "UITestCaseStatusInfo_EndIsEmpty",
                "UITestCaseStatusInfo_EndLessThenStart",
                "HealthChecks_NameIsEmpty",
                "HealthChecks_NameMaxLength",
                "HealthChecks_UriIsEmpty",
                "UserFiles_FileNameIsEmpty",
                "UserFiles_FileSizeIsNegative",
                "UserFiles_FileNameMaxLength",
                "PageIndexLessThan0",
                "PageSizeLessThen1",
                "DeletionFailed"
              ],
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "GetUIElementDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "FindByMethod": {
            "enum": [
              "ById",
              "ByClass",
              "ByXPath"
            ],
            "type": "string"
          },
          "FindBy": {
            "type": "string",
            "nullable": true
          },
          "PageId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GetUIElementDtoPaginationResult": {
        "type": "object",
        "properties": {
          "Count": {
            "type": "integer",
            "format": "int32"
          },
          "Data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetUIElementDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetUIPageDetailsDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Page": {
            "$ref": "#/components/schemas/GetUIElementDtoPaginationResult"
          }
        },
        "additionalProperties": false
      },
      "GetUIPageDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetUIPageDtoPaginationResult": {
        "type": "object",
        "properties": {
          "Count": {
            "type": "integer",
            "format": "int32"
          },
          "Data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetUIPageDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PostResponse": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "required": [
          "TypeDiscriminator"
        ],
        "type": "object",
        "properties": {
          "TypeDiscriminator": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { },
        "discriminator": {
          "propertyName": "TypeDiscriminator"
        }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          }
        ],
        "properties": {
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}